test(range): migrate scroll-target
